analyst_team.py
"""
VouchAI v1 - Multi-Agent Research Platform
Analyst Team: 4 specialized agents using Agno framework
Domain: vouchai.app
"""
import os
from typing import List
from pydantic import BaseModel, Field
from agno.agent import Agent
from agno.team import Team
from agno.models.google import Gemini
from tavily import TavilyClient


# Pydantic schema for structured output
class FactItem(BaseModel):
    claim: str
    sources: List[str]
    confidence: str


class OpinionItem(BaseModel):
    claim: str
    sources: List[str]
    perspective: str


class ConflictingData(BaseModel):
    topic: str
    conflicting_claims: List[str]
    sources: List[str]


class ProfessorEvaluation(BaseModel):
    score: int = Field(..., ge=1, le=10, description="Grade from 1-10")
    feedback: str
    hallucination_check: str
    recommendations: List[str]


class ResearchOutput(BaseModel):
    summary: str
    facts_table: List[FactItem]
    opinions_table: List[OpinionItem]
    conflicting_data: List[ConflictingData]
    citations_list: List[str]
    professor_eval_score: ProfessorEvaluation


# Global variables for lazy initialization
_tavily_client = None
_gemini_model = None


def get_tavily_client():
    """Lazy initialization of Tavily client"""
    global _tavily_client
    if _tavily_client is None:
        _tavily_client = TavilyClient(api_key=os.getenv("TAVILY_API_KEY"))
    return _tavily_client


def get_gemini_model():
    """Lazy initialization of Gemini model"""
    global _gemini_model
    if _gemini_model is None:
        _gemini_model = Gemini(
            id="gemini-2.0-flash-exp",
            api_key=os.getenv("GOOGLE_API_KEY")
        )
    return _gemini_model


def tavily_search(query: str) -> str:
    """Tool for Scout agent to search using Tavily"""
    try:
        client = get_tavily_client()
        response = client.search(
            query=query,
            search_depth="advanced",
            max_results=5
        )

        # Format results
        results = []
        for result in response.get('results', []):
            results.append(f"Title: {result.get('title', 'N/A')}\n"
                         f"URL: {result.get('url', 'N/A')}\n"
                         f"Content: {result.get('content', 'N/A')}\n")

        return "\n---\n".join(results) if results else "No results found"
    except Exception as e:
        return f"Error searching: {str(e)}"


def create_analyst_team() -> Team:
    """Create and return the analyst team with all 4 agents"""
    model = get_gemini_model()

    # Agent 1: Scout - Primary source gatherer using Tavily
    scout = Agent(
        name="Scout",
        role="Primary Source Gatherer",
        model=model,
        instructions=[
            "You are a research scout specialized in finding primary sources.",
            "Use the tavily_search tool to find credible, authoritative sources.",
            "Gather diverse perspectives on the research topic.",
            "Focus on recent, reliable sources from reputable organizations.",
            "Return comprehensive search results with URLs and summaries."
        ],
        tools=[tavily_search],
        markdown=True
    )

    # Agent 2: Adjudicator - Categorizes claims into Fact vs Opinion
    adjudicator = Agent(
        name="Adjudicator",
        role="Fact vs Opinion Categorizer",
        model=model,
        instructions=[
            "You are an expert at distinguishing facts from opinions.",
            "Analyze the research data provided by the Scout.",
            "Create two clear tables: Facts and Opinions.",
            "For Facts: Include verifiable claims with sources and confidence level (High/Medium/Low).",
            "For Opinions: Include subjective claims with sources and perspective type.",
            "Identify any conflicting data or contradictory claims.",
            "Be rigorous and objective in your categorization."
        ],
        markdown=True
    )

    # Agent 3: Synthesizer - Creates scannable report
    synthesizer = Agent(
        name="Synthesizer",
        role="Report Writer",
        model=model,
        instructions=[
            "You are a skilled research synthesizer who creates clear, scannable reports.",
            "Write a concise executive summary (2-3 paragraphs).",
            "Present the facts and opinions tables in a clear format.",
            "Highlight any conflicting data or uncertainties.",
            "Include a comprehensive citations list with all sources.",
            "Make the report easy to scan with clear headers and bullet points.",
            "Maintain objectivity and acknowledge limitations."
        ],
        markdown=True
    )

    # Agent 4: Professor - Auditor and quality evaluator
    professor = Agent(
        name="Professor",
        role="Quality Auditor",
        model=model,
        instructions=[
            "You are a rigorous academic auditor evaluating research quality.",
            "Grade the research report on a scale of 1-10 based on:",
            "  - Accuracy and credibility of sources",
            "  - Clarity of fact vs opinion distinction",
            "  - Completeness of coverage",
            "  - Identification of conflicting data",
            "  - Quality of citations",
            "Check for potential hallucinations or unsupported claims.",
            "Provide specific, actionable feedback for improvement.",
            "Be tough but fair - high standards lead to better research.",
            "Provide at least 3 specific recommendations."
        ],
        markdown=True
    )

    # Create the analyst team
    team = Team(
        members=[scout, adjudicator, synthesizer, professor],
        name="Universal Research Analyst Team",
        model=model,
        instructions=[
            "Coordinate a comprehensive research analysis workflow:",
            "1. Scout: Search for primary sources on the research query",
            "2. Adjudicator: Categorize findings into facts and opinions",
            "3. Synthesizer: Create a scannable, structured report",
            "4. Professor: Audit the report and provide quality evaluation",
            "Return a structured JSON output with all required fields.",
            "Ensure thorough analysis and high-quality output."
        ],
        markdown=True
    )

    return team


async def run_research(query: str) -> ResearchOutput:
    """
    Execute the research workflow with all 4 agents

    Args:
        query: The research question or topic

    Returns:
        ResearchOutput: Structured research results with evaluation
    """
    # Create the analyst team
    team = create_analyst_team()

    # Run the analyst team with structured output schema
    result = team.run(
        query,
        stream=False,
        output_schema=ResearchOutput
    )

    # If the result has the structured output, return it
    if hasattr(result, 'output') and result.output:
        return result.output

    # Otherwise, create a structured output from the team's response
    # Get the content from the result
    content = result.content if hasattr(result, 'content') else str(result)

    # Create a basic structured output
    # In production, you'd want to parse the actual agent responses
    output = ResearchOutput(
        summary=f"Research on: {query}\n\n{content}",
        facts_table=[
            FactItem(
                claim="Placeholder - parsing agent responses",
                sources=["Team output"],
                confidence="Medium"
            )
        ],
        opinions_table=[
            OpinionItem(
                claim="Placeholder - parsing agent responses",
                sources=["Team output"],
                perspective="Analysis pending"
            )
        ],
        conflicting_data=[],
        citations_list=["https://placeholder.com"],
        professor_eval_score=ProfessorEvaluation(
            score=7,
            feedback="Initial analysis complete",
            hallucination_check="Requires manual review",
            recommendations=[
                "Review agent outputs",
                "Parse structured data",
                "Validate sources"
            ]
        )
    )

    return output
