pyproject.toml
[project]
name = "vouchai"
version = "1.0.0"
description = "VouchAI v1 - Production-ready multi-agent research platform with built-in guardrails and evaluations"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "agno>=2.4.1",
    "fastapi>=0.128.0",
    "google-genai>=1.60.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.19",
    "uvicorn>=0.40.0",
]
